requirements.txt
python-telegram-bot==20.7
openai==1.3.5
youtube-transcript-api==0.6.1
yt-dlp==2023.11.16
whisper==1.1.10
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1 